incr

+++ examples/neopixel_demo.py
@@ -49,7 +49,7 @@ print(ALT_BTF)
 ALT2_BTF = []
 map_x = xsize
 map_y = ysize
-ylist = list(range(map_y/2))
+ylist = list(range(int(map_y/2)))
 for y in ylist:
     xlist = list(range(y*(map_x*2),y*(map_x*2)+map_x*2))
     for x in xlist:
